+++ packages/graphql-toolbox/src/constants.ts
@@ -22,7 +22,7 @@ export const LOCAL_STATE_OPTIONS = "neo4j.graphql.options";
 export const LOCAL_STATE_TYPE_LAST_QUERY = "neo4j.graphql.lastQuery";
 export const LOCAL_STATE_TYPE_LAST_PARAMS = "neo4j.graphql.lastParams";
 export const LOCAL_STATE_LOGIN = "neo4j.graphql.login";
-export const LOCAL_STATE_DEBUG = "neo4j.graphql.debug";
+export const LOCAL_STATE_ENABLE_DEBUG = "neo4j.graphql.debug";
 export const LOCAL_STATE_CHECK_CONSTRAINT = "neo4j.graphql.checkConstraint";
 export const LOCAL_STATE_CREATE_CONSTRAINT = "neo4j.graphql.createConstraint";
 export const LOCAL_STATE_EDITOR_THEME = "neo4j.graphql.editorTheme";
