+++ packages/graphql-toolbox/src/modules/EditorView/Editor.tsx
@@ -35,13 +35,14 @@ import {
 } from "../../constants";
 import { Grid } from "./grid/Grid";
 import { formatCode, safeParse, ParserOptions } from "./utils";
-import { Extension } from "../Filename";
-import { ViewSelectorComponent } from "../ViewSelectorComponent";
+import { Extension } from "../../components/Filename";
+import { ViewSelectorComponent } from "../../components/ViewSelectorComponent";
 import { SettingsContext } from "../../contexts/settings";
 import { Theme, ThemeContext } from "../../contexts/theme";
-import { AppSettings } from "../drawers/AppSettings";
-import { ProTooltip } from "../ProTooltip";
-import { HelpDrawer } from "../drawers/HelpDrawer";
+import { AppSettings } from "../Drawers/AppSettings";
+import { ProTooltip } from "../../components/ProTooltip";
+import { HelpDrawer } from "../Drawers/HelpDrawer";
+import { Storage } from "../../utils/storage";
 
 const DEBOUNCE_TIMEOUT = 500;
 export interface Props {
@@ -62,7 +63,7 @@ export const Editor = (props: Props) => {
 
     const debouncedSave = useCallback(
         debounce((key, value) => {
-            localStorage.setItem(key, value);
+            Storage.store(key, value);
         }, DEBOUNCE_TIMEOUT),
         []
     );
@@ -100,8 +101,8 @@ export const Editor = (props: Props) => {
     );
 
     useEffect(() => {
-        const initQuery = JSON.parse(localStorage.getItem(LOCAL_STATE_TYPE_LAST_QUERY) as string) || DEFAULT_QUERY;
-        const initParams = JSON.parse(localStorage.getItem(LOCAL_STATE_TYPE_LAST_PARAMS) as string) || "";
+        const initQuery = Storage.retrieveJSON(LOCAL_STATE_TYPE_LAST_QUERY) || DEFAULT_QUERY;
+        const initParams = Storage.retrieveJSON(LOCAL_STATE_TYPE_LAST_PARAMS) || "";
         setInitialLoad(true);
         setQuery(initQuery);
         setVariableValues(initParams);
