+++ packages/graphql-toolbox/src/modules/EditorView/docexplorer/SearchBox.tsx
